% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_n.R
\name{adjust_n}
\alias{adjust_n}
\title{Adjust number of samples}
\usage{
adjust_n(dates, freq, n_size = 8)
}
\arguments{
\item{dates}{A vector of dates.}

\item{freq}{A character vector. May be "week", "month", "quarter", or "year".}

\item{n_size}{Integer of the number to be adjusted.}
}
\value{
An integer
}
\description{
Adjust number of samples
}
\keyword{internal}
