% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_seasons.R
\name{calc_seasons}
\alias{calc_seasons}
\title{Calculate seasons from dates}
\usage{
calc_seasons(dates, n_seasons = 4, season_start = 10)
}
\arguments{
\item{dates}{A vector of dates of 'Date' class.}

\item{n_seasons}{The number of seasons in a year. Must be a factor of 12. Default is 4.}

\item{season_start}{The starting month of the first season, specified as an integer from 1 to
12. Default is 10 (October).}
}
\value{
A vector of named integers representing the season for each date in \code{dates}.
}
\description{
Assign seasons to a vector of dates based on the specified number of seasons and the starting
month of the first season.
}
\examples{
date_vec <- as.Date(c("2020-01-01", "2020-03-01", "2020-06-01", "2020-09-01", "2020-12-01"))
streamsampler:::calc_seasons(date_vec)

date_vec <- as.Date(c("2020-01-01", "2020-06-01", "2020-07-01", "2020-12-01"))
streamsampler:::calc_seasons(date_vec, n_seasons = 2)

date_vec <- as.Date(c("2020-01-01", "2020-03-01", "2020-07-01", "2020-09-01", "2020-12-01"))
streamsampler:::calc_seasons(date_vec, season_start = 7)

date_vec <- seq.Date(as.Date("2020-01-01"), as.Date("2020-12-31"), by = "month")
streamsampler:::calc_seasons(date_vec)

date_vec <- as.Date(c("2020-12-31", "2021-01-01"))
streamsampler:::calc_seasons(date_vec)

date_vec <- as.Date(c("2020-01-01", "2020-04-01", "2020-08-01", "2020-12-01"))
streamsampler:::calc_seasons(date_vec, n_seasons = 3, season_start = 4)
}
\keyword{internal}
