% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_year.R
\name{center_year}
\alias{center_year}
\title{Find middle year in a sequence of years}
\usage{
center_year(years, start_year, end_year, half_win = 2)
}
\arguments{
\item{years}{Vector of a sequence of years}

\item{start_year}{The first year of the record}

\item{end_year}{The last year of the record}

\item{half_win}{The half-window size, specifying the range of years to be maintained,
particularly from the start and end years.}
}
\value{
Vector of the center years
}
\description{
Find middle year in a sequence of years
}
\examples{
years <- c(2005, 2010, 2015)
start_year <- 2000
end_year <- 2020
half_win <- 5
streamsampler:::center_year(years, start_year, end_year, half_win)

years <- c(1995, 2000, 2002)
start_year <- 2000
end_year <- 2020
half_win <- 5
streamsampler:::center_year(years, start_year, end_year, half_win)

years <- c(2018, 2021, 2025)
start_year <- 2000
end_year <- 2020
half_win <- 5
streamsampler:::center_year(years, start_year, end_year, half_win)

years <- c(2000, 2020)
start_year <- 2000
end_year <- 2020
half_win <- 5
streamsampler:::center_year(years, start_year, end_year, half_win)

years <- 2010
start_year <- 2000
end_year <- 2020
half_win <- 5
streamsampler:::center_year(years, start_year, end_year, half_win)
}
\keyword{internal}
