% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_eval_dates_inputs}
\alias{check_eval_dates_inputs}
\title{Check inputs to eval_dates()}
\usage{
check_eval_dates_inputs(dates, rec_start, rec_end, by)
}
\arguments{
\item{dates}{A vector of dates to be evaluated. Must be of class 'Date'.}

\item{rec_start}{The start date of the recording period. Must be of class 'Date'.}

\item{rec_end}{The end date of the recording period. Must be of class 'Date'.}

\item{by}{A character string specifying the time period for aggregation. One of "day", "week",
"month", "quarter", or "year". Default is "day".}
}
\value{
A TRUE/FALSE value with a "msg" attribute
}
\description{
Check inputs to eval_dates()
}
\keyword{internal}
