% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_qw_inputs}
\alias{check_qw_inputs}
\title{Check inputs to qw_stats()}
\usage{
check_qw_inputs(dates, values, rec_start, rec_end, by)
}
\arguments{
\item{dates}{A vector of dates to be evaluated. Must be of class \code{Date}.}

\item{values}{A numeric vector to be evaluated. Must be of class \code{numeric}.}

\item{rec_start}{The start date of the recording period. Must be of class \code{Date}.}

\item{rec_end}{The end date of the recording period. Must be of class \code{Date}.}

\item{by}{The time interval to use for the evaluation. One of "day", "week", "month", "quarter",
or "year". Default is "day".}
}
\value{
A TRUE/FALSE value with a "msg" attribute
}
\description{
Validates input values to qw_stats() and returns a message for errors
}
\keyword{internal}
