% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_rollmean_date_inputs}
\alias{check_rollmean_date_inputs}
\title{Check inputs to rollmean_date()}
\usage{
check_rollmean_date_inputs(dates, values, look_behind, look_ahead, look_units)
}
\arguments{
\item{dates}{A vector of dates of 'Date' class.}

\item{values}{Numeric values. The values must be in correspondence with \code{dates}, meaning the
\emph{i}th element in \code{values} must correspond to the \emph{i}th date in \code{dates}.}

\item{look_behind, look_ahead}{The number of \code{look_units} before and after the center date to
include in the sliding window.}

\item{look_units}{One of "days", "weeks", or "months". The units to give \code{look_ahead} and
\code{look_behind}.}
}
\value{
A TRUE/FALSE value with a "msg" attribute
}
\description{
Check inputs to rollmean_date()
}
\keyword{internal}
