% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_subsample_routine_inputs}
\alias{check_subsample_routine_inputs}
\title{Check inputs to subsample_routine()}
\usage{
check_subsample_routine_inputs(dates, values, day, freq)
}
\arguments{
\item{dates}{A vector of dates of 'Date' class.}

\item{values}{Numeric values. The values must be in correspondence with \code{dates}, meaning the
\emph{i}th element in \code{values} must correspond to the \emph{i}th date in \code{dates}.}

\item{day}{An integer specifying the day of the specified \code{freq}uency. See 'Details'.}

\item{freq}{A character string indicating the frequency of selection. Must be one of "day",
"week", or "month". See 'Details'.}
}
\value{
A TRUE/FALSE value with a "msg" attribute
}
\description{
Check inputs to subsample_routine()
}
\keyword{internal}
