% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_summarize_seasons_inputs}
\alias{check_summarize_seasons_inputs}
\title{Check inputs to summarize_seasons()}
\usage{
check_summarize_seasons_inputs(dates, values, season_start, n_seasons)
}
\arguments{
\item{dates}{A vector of dates of 'Date' class.}

\item{values}{Numeric values. The values must be in correspondence with \code{dates}, meaning the
\emph{i}th element in \code{values} must correspond to the \emph{i}th date in \code{dates}.}

\item{season_start}{The starting month of the first season, specified as an integer from 1 to
12. Default is 10 (October).}

\item{n_seasons}{The number of seasons in a year. Must be a factor of 12. Default is 4.}
}
\value{
A TRUE/FALSE value with a "msg" attribute
}
\description{
Check inputs to summarize_seasons()
}
\keyword{internal}
