% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_thresholds_inputs}
\alias{check_thresholds_inputs}
\title{Check inputs to thresholds()}
\usage{
check_thresholds_inputs(
  dates,
  values,
  season_start,
  n_seasons,
  half_win,
  threshold
)
}
\arguments{
\item{dates}{A vector of dates of 'Date' class.}

\item{values}{Numeric values. The values must be in correspondence with \code{dates}, meaning the
\emph{i}th element in \code{values} must correspond to the \emph{i}th date in \code{dates}.}

\item{season_start}{The starting month of the first season, specified as an integer from 1 to
12. Default is 10 (October).}

\item{n_seasons}{The number of seasons in a year. Must be a factor of 12. Default is 4.}

\item{half_win}{The half width of the window of years to group \code{thresh_ref} by. See 'Details'.}

\item{threshold}{The quantile of \code{thresh_ref} above which \code{values} is sampled \code{n_et_samples}
times per year.}
}
\value{
A TRUE/FALSE value with a "msg" attribute
}
\description{
Validates input values to thresholds() and returns a message for errors
}
\keyword{internal}
