% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamdat.R
\docType{data}
\name{streamdat}
\alias{streamdat}
\title{Stream Discharge and Specific Conductivity}
\format{
A data frame with 5,844 rows and 3 columns:
\describe{
\item{date}{Date of measurement}
\item{q}{Discharge in cubic feet per second}
\item{sc}{Specific conductivity in microsiemens per centimeter at 25 degrees Celsius}
}
}
\source{
\url{https://waterdata.usgs.gov/monitoring-location/01481500/#parameterCode=00060&period=P7D&showMedian=false}
}
\usage{
streamdat
}
\description{
Approximately 15 years of daily average discharge and specific conductivity
of the Brandywine Creek at Wilmington, DE - stream gage number 01481500.
}
\keyword{datasets}
