% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamsampler-package.R
\docType{package}
\name{streamsampler-package}
\alias{streamsampler}
\alias{streamsampler-package}
\title{streamsampler: Characterize and Subsample Stream Data}
\description{
A series of functions that characterizes daily stream discharge and water quality data and subsamples the water quality data. Provide dates, discharge, and water quality measurements and `streamsampler` can find gaps, get summary statistics, and subsample according to common stream sampling protocols.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://code.usgs.gov/water/streamsampler}
  \item Report bugs at \url{https://code.usgs.gov/water/streamsampler/-/issues}
}

}
\author{
\strong{Maintainer}: Kyle Hurley \email{khurley@usgs.gov} (\href{https://orcid.org/0009-0002-6469-3769}{ORCID})

Authors:
\itemize{
  \item Jeff Chanat \email{jchanat@usgs.gov} (\href{https://orcid.org/0000-0002-3629-7307}{ORCID})
}

}
\keyword{internal}
