% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample_et}
\alias{subsample_et}
\title{Subsample exceeds-threshold data}
\usage{
subsample_et(group, n_et_samples = 8, seed = 123)
}
\arguments{
\item{group}{Data frame to be subsampled. Must include columns named "date" ('Date'), "value"
('numeric'), "excd_thresh" (logical), and "selection_type" (character).}

\item{n_et_samples}{Integer of the number of yearly exceeds-threshold samples to be selected
from \code{values}. See 'Details'.}

\item{seed}{An integer which determines the state for the random number generator. Ensures
random sampling is reproducible.}
}
\description{
Samples values that exceed the threshold
}
\keyword{internal}
